requirements.txt
applescript==2020.4.15
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
detect==2019.4.13
idna==2.9
only==2019.4.13
Pillow==7.1.2
public==2019.4.13
requests==2.23.0
soupsieve==2.0.1
Unidecode==1.1.1
urllib3==1.25.9
